+++ src/mnemocards_anki/__init__.py
@@ -1,5 +1,6 @@
 from mnemocards_anki.configure import Configure as Configure
 from mnemocards_anki.package import Package as Package
+from mnemocards_anki.pronounce import Pronounce as Pronounce
 from mnemocards_anki.simple_vocabulary import (
     SimpleVocabularyNoteType as SimpleVocabularyNoteType,
 )